Authorisation des bien immo et des adresses

+++ Controllers/AdressesController.cs
@@ -8,6 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using ASimmo.Data;
 using ASimmo.Models;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace ASimmo.Controllers
 {
@@ -21,9 +22,21 @@ namespace ASimmo.Controllers
         }
 
         // GET: Adresses
+        [Authorize(Roles = "Admin, Agent")]
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Adresses.ToListAsync());
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                var applicationDbContext = _context.Adresses.Include(c => c.Promoteur);
+                return View(await applicationDbContext.ToListAsync());
+            }
+            else if (this.HttpContext.User.IsInRole("Agent"))
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var applicationDbContext = _context.Adresses.Include(c => c.Promoteur).Where(p => p.Promoteur.UserId == _userId);
+                return View(await applicationDbContext.ToListAsync());
+            }
+            return new ForbidResult();
         }
 
         // GET: Adresses/Details/5
@@ -48,6 +61,15 @@ namespace ASimmo.Controllers
         [Authorize(Roles = "Admin, Agent")]
         public IActionResult Create()
         {
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs, "PromoteurId", "Libelle");
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs.Where(p => p.UserId == _userId), "PromoteurId", "Libelle");
+            }
             return View();
         }
 
@@ -57,13 +79,29 @@ namespace ASimmo.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize(Roles = "Admin, Agent")]
-        public async Task<IActionResult> Create([Bind("AdresseId,Quartier,CodePostal,Ville,AdressePostale")] Adresse adresse)
+        public async Task<IActionResult> Create([Bind("AdresseId,Quartier,CodePostal,Ville,AdressePostale,PromoteurId")] Adresse adresse)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(adresse);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (this.HttpContext.User.IsInRole("Admin") || IsOwner(adresse.PromoteurId))
+                {
+                    _context.Add(adresse);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    return new ForbidResult();
+                }
+            }
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs, "PromoteurId", "Libelle");
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs.Where(p => p.UserId == _userId), "PromoteurId", "Libelle");
             }
             return View(adresse);
         }
@@ -82,6 +120,15 @@ namespace ASimmo.Controllers
             {
                 return NotFound();
             }
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs, "PromoteurId", "Libelle");
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs.Where(p => p.UserId == _userId), "PromoteurId", "Libelle");
+            }
             return View(adresse);
         }
 
@@ -91,7 +138,7 @@ namespace ASimmo.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize(Roles = "Admin, Agent")]
-        public async Task<IActionResult> Edit(int id, [Bind("AdresseId,Quartier,CodePostal,Ville,AdressePostale")] Adresse adresse)
+        public async Task<IActionResult> Edit(int id, [Bind("AdresseId,Quartier,CodePostal,Ville,AdressePostale,PromoteurId")] Adresse adresse)
         {
             if (id != adresse.AdresseId)
             {
@@ -102,8 +149,16 @@ namespace ASimmo.Controllers
             {
                 try
                 {
-                    _context.Update(adresse);
-                    await _context.SaveChangesAsync();
+                    if (this.HttpContext.User.IsInRole("Admin") || IsOwner(adresse.PromoteurId))
+                    {
+                        _context.Update(adresse);
+                        await _context.SaveChangesAsync();
+                    }
+                    else
+                    {
+                        return new ForbidResult();
+                    }
+
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -117,6 +172,17 @@ namespace ASimmo.Controllers
                     }
                 }
                 return RedirectToAction(nameof(Index));
+            }
+
+
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs, "PromoteurId", "Libelle");
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["PromoteurId"] = new SelectList(_context.Promoteurs.Where(p => p.UserId == _userId), "PromoteurId", "Libelle");
             }
             return View(adresse);
         }
@@ -136,8 +202,15 @@ namespace ASimmo.Controllers
             {
                 return NotFound();
             }
+            if (this.HttpContext.User.IsInRole("Admin") || IsOwner(adresse.PromoteurId))
+            {
+                return View(adresse);
+            }
+            else
+            {
+                return new ForbidResult();
+            }
 
-            return View(adresse);
         }
 
         // POST: Adresses/Delete/5
@@ -147,14 +220,33 @@ namespace ASimmo.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var adresse = await _context.Adresses.FindAsync(id);
-            _context.Adresses.Remove(adresse);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (this.HttpContext.User.IsInRole("Admin") || IsOwner(adresse.PromoteurId))
+            {
+                _context.Adresses.Remove(adresse);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return new ForbidResult();
+            }
         }
 
         private bool AdresseExists(int id)
         {
             return _context.Adresses.Any(e => e.AdresseId == id);
+        }
+
+        private bool IsOwner(int pid)
+        {
+            var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var promoteur = _context.Promoteurs.Include(p => p.User).FirstOrDefaultAsync(p => p.PromoteurId == pid).GetAwaiter().GetResult();
+            if (promoteur.UserId == _userId)
+            {
+                return true;
+            }
+
+            return false;
         }
     }
 }
