+++ Controllers/BienImmosController.cs
@@ -8,26 +8,38 @@ using Microsoft.EntityFrameworkCore;
 using ASimmo.Data;
 using ASimmo.Models;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace ASimmo.Controllers
 {
-    public class BienImmoesController : Controller
+    public class BienImmosController : Controller
     {
         private readonly ApplicationDbContext _context;
 
-        public BienImmoesController(ApplicationDbContext context)
+        public BienImmosController(ApplicationDbContext context)
         {
             _context = context;
         }
 
-        // GET: BienImmoes
+        // GET: BienImmos
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.BiensImmos.Include(b => b.Adresse).Include(b => b.Classification).Include(b => b.Type);
-            return View(await applicationDbContext.ToListAsync());
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                var applicationDbContext = _context.BiensImmos.Include(b => b.Adresse).Include(b => b.Classification).Include(b => b.Type);
+                return View(await applicationDbContext.ToListAsync());
+            }
+            else if (this.HttpContext.User.IsInRole("Agent"))
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                var applicationDbContext = _context.BiensImmos.Include(b => b.Adresse).Include(b => b.Classification).Include(b => b.Type).Include(b=> b.Classification.Promoteur).Where(b => b.Classification.Promoteur.UserId == _userId);
+                return View( await applicationDbContext.ToListAsync());
+            }
+            return new ForbidResult();
+            
         }
 
-        // GET: BienImmoes/Details/5
+        // GET: BienImmos/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -52,9 +64,20 @@ namespace ASimmo.Controllers
         [Authorize(Roles = "Admin, Agent")]
         public IActionResult Create()
         {
-            ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdresseId");
-            ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "ClassificationId");
-            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "TypeBienImmoId");
+            if(this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdressePostale");
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "Libelle");
+
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["AdresseId"] = new SelectList(_context.Adresses.Include(a=> a.Promoteur).Where(a => a.Promoteur.UserId == _userId), "AdresseId", "AdressePostale");
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications.Include(c => c.Promoteur).Where(c => c.Promoteur.UserId == _userId), "ClassificationId", "Libelle");
+            }
+
+            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "Libelle");
             return View();
         }
 
@@ -72,9 +95,20 @@ namespace ASimmo.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdresseId", bienImmo.AdresseId);
-            ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "ClassificationId", bienImmo.ClassificationId);
-            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "TypeBienImmoId", bienImmo.TypeId);
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "Libelle", bienImmo.ClassificationId);
+
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["AdresseId"] = new SelectList(_context.Adresses.Include(a => a.Promoteur).Where(a => a.Promoteur.UserId == _userId), "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications.Include(c => c.Promoteur).Where(c => c.Promoteur.UserId == _userId), "ClassificationId", "Libelle", bienImmo.ClassificationId);
+            }
+
+            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "Libelle", bienImmo.TypeId);
             return View(bienImmo);
         }
 
@@ -92,9 +126,20 @@ namespace ASimmo.Controllers
             {
                 return NotFound();
             }
-            ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdresseId", bienImmo.AdresseId);
-            ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "ClassificationId", bienImmo.ClassificationId);
-            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "TypeBienImmoId", bienImmo.TypeId);
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "Libelle", bienImmo.ClassificationId);
+
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["AdresseId"] = new SelectList(_context.Adresses.Include(a => a.Promoteur).Where(a => a.Promoteur.UserId == _userId), "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications.Include(c => c.Promoteur).Where(c => c.Promoteur.UserId == _userId), "ClassificationId", "Libelle", bienImmo.ClassificationId);
+            }
+
+            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "Libelle", bienImmo.TypeId);
             return View(bienImmo);
         }
 
@@ -115,8 +160,15 @@ namespace ASimmo.Controllers
             {
                 try
                 {
-                    _context.Update(bienImmo);
-                    await _context.SaveChangesAsync();
+                    if (this.HttpContext.User.IsInRole("Admin") || IsOwner(bienImmo.ClassificationId))
+                    {
+                        _context.Update(bienImmo);
+                        await _context.SaveChangesAsync();
+                    }
+                    else
+                    {
+                        return new ForbidResult();
+                    }
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -131,9 +183,20 @@ namespace ASimmo.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdresseId", bienImmo.AdresseId);
-            ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "ClassificationId", bienImmo.ClassificationId);
-            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "TypeBienImmoId", bienImmo.TypeId);
+
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                ViewData["AdresseId"] = new SelectList(_context.Adresses, "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications, "ClassificationId", "Libelle", bienImmo.ClassificationId);
+
+            }
+            else
+            {
+                var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                ViewData["AdresseId"] = new SelectList(_context.Adresses.Include(a => a.Promoteur).Where(a => a.Promoteur.UserId == _userId), "AdresseId", "AdressePostale", bienImmo.AdresseId);
+                ViewData["ClassificationId"] = new SelectList(_context.Classifications.Where(c => IsOwner(c.PromoteurId)), "ClassificationId", "Libelle", bienImmo.ClassificationId);
+            }
+            ViewData["TypeId"] = new SelectList(_context.TypesBiensImmos, "TypeBienImmoId", "Libelle", bienImmo.TypeId);
             return View(bienImmo);
         }
 
@@ -155,8 +218,13 @@ namespace ASimmo.Controllers
             {
                 return NotFound();
             }
-
-            return View(bienImmo);
+            if (this.HttpContext.User.IsInRole("Admin") || IsOwner(bienImmo.ClassificationId))
+            {
+                return View(bienImmo);
+            } else
+            {
+                return new ForbidResult();
+            }
         }
 
         // POST: BienImmoes/Delete/5
@@ -166,14 +234,34 @@ namespace ASimmo.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var bienImmo = await _context.BiensImmos.FindAsync(id);
-            _context.BiensImmos.Remove(bienImmo);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (this.HttpContext.User.IsInRole("Admin") || IsOwner(bienImmo.ClassificationId))
+            {
+                
+                _context.BiensImmos.Remove(bienImmo);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return new ForbidResult();
+            }
         }
 
         private bool BienImmoExists(int id)
         {
             return _context.BiensImmos.Any(e => e.BienImmoId == id);
+        }
+
+        private bool IsOwner(int cid)
+        {
+            var _userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var _c = _context.Classifications.Include(c=> c.Promoteur).Include(c => c.Promoteur.User).Where(c=> c.Promoteur.User.Id == _userId).FirstOrDefault(c => c.ClassificationId == cid);
+            if (_c != null)
+            {
+                return true;
+            }
+            return false;
         }
     }
 }
